+++ nisystemlink/clients/spec/__init__.py
@@ -1,0 +1,3 @@
+from ._spec_client import SpecClient
+
+# flake8: noqa